tried one more time. failed
tried one more time to get the animation going. proplem is, when I input the GameObject into the preview tab, it sinks halfway into the ground and no animation plays. it is not even T-posing.
tried again and it suddenly worked (for idle at least) but the walking animation is not hooked up.